third_party/blink/web_tests/fast/workers/chromium/worker-messageport-gc-expected.txt

Test that workers stay reachable via message ports. Should print "DONE" when done.

PASS: Worker is reachable.
DONE
